Port to GTK+ 3 and support Python 3 (#4)
Support for both was added by using pygtkcompat. Minimal changes
were required to allow this.

The source code was also run through a formatter to correct the
heavy mix of tabs and spaces. I have no idea how it even ran with
such inconsistent indentation.

This fixes the issue that the equalizer only worked with python 2.7.

[thanks to @kabili207 ]