redesign: use gtkbuilder ui file
Add ui template file for the Equalizer ApplicationWindow and make use of
it trough @Gtk.Template.

The ui file is compiled into a GResource which needs to be sourced
before the Equalizer class file gets imported.

Most of the not dynamically created widgets have beend moved from the
python source to the template. Buttons signals are also mapped in the ui
file. The corresponding callback functions are marked with
@Gtk.Template.Callback().