redesign: change EQ Enabled to a headerbar switch
Change the EQ Enabled CheckBox into a Switch and move it into the top
right of the headerbar.

Represent the state of EQ Enabled with a statefull action and connect to
its state-changed signal instead of connecting to the button directly.

Remove the now redundant title suffix of enabled and disabled as it is
now directly visible trough the switch.

Remove occurences of the unused variable windowtitle.