redesign: move save button to headerbar
This commit moves the save button to the headerbar.
Its label has been replaced with a symbolic icon.

A GAction has been added for saving a preset and is used by the button.

In additon the save button is now preset sensitive and will only be
enabled when it is actually possible to save a preset.
It is possible to save a preset if it has a not yet existing and not
empty preset name.

This is the same behavior as with the old design but instead of failing
silently we now do not allow the save action to be used in those invalid
cases.