redesign: autoapply changed scales
Changed scale values are now automatically applied after a change if
there was no other scale value change within 500ms. If there was such a
change the timer is reset to 500ms.

This approach for automatically applying changes was chosen because
applying the change after every change to the scale would cause the
application to hang since the equalizer module has to reloaded.

In addition it is not easily possible to determine when a scroll event
on the scale has ended other then having no change to the value for some
time.

The now redundant apply settings button has been removed.