preamp: remove preamp from ui code
The preamp value was currently only read from settings and not from the
presets itself. It still was written to the preset files but never read.

A scale for the preamp value was added to the user interface but never
shown.

The preamp value is read in pulse-audio-equalizer-ladspa script but
never used.

This commit removes the preamp user interface part for a cleaner code.

In addition no preamp value will be written to new preset but an empty
string.

Fixes #8