refactor: move frequency labels to their own class
This allows to correctly format the frequency labels without the help of
globals. The FrequencyLabel has one method set_frequency which takes a
frequency string as parameter and will format the label correctly as
with the old FormatLabels() function.

The frequency can also be set in the objects constructor with the
frequency keyword parameter.