meson: add python purelib path option
This commit adds the option purelib which is the python purelib path
relative to the prefix. This option is needed for distributions that
have not set the purelib path correctly in sysconfig. e.g. Debian/Ubuntu
which use dist-packages subfolder rather then site-packages.

For example for Ubuntu 18.10 use -Dpurelib=lib/python3.6/dist-packages

Signed-off-by: Gabriel Rauter <rauter.gabriel@gmail.com>